initial commit
This repo holds most of my vim configuration in the form of a plugin.
I'll be using Vim 8's package feature to pack this with the other
plugins it depends on. It's loosely coupled to these plugins, mostly by
<Plug> mappings (though there is also a direct call to fugitive#head()
in the statusline). It also sets a number of global variables that
affect plugin behavior, so those plugins are loosely dependent on this
one.