fix initialization logic
Vim doesn't seem to load the contents of after/ directories in packages.
I had hoped to put this colorscheme setting in after/ to ensure the coot
plugin has been loaded before I call on one of its colorschemes. I'm
going to try calling it earlier. If coot has been added to the
runtimepath, it should find the colorscheme.

Also, remove a bunch of &cp checks, since I'm really not designing this
config for that kind of extreme backwards compatibility.