mline: move to after/ so fugitive can load
In order to losen the dependency on fugitive (so I can break mline back
out into its own plugin), make the branch section of statusline
conditional on the fugitive#head function existing. That also means we
have to load this later to give fugitive a chance to load.

I wonder though, will this work? Since fugitive#head is autoloaded, it's
likely not going to exist yet. I may have to actually call it.