reorg autocmds, extract cursor
I'm finding that grouping autocommands into a common file is a bad idea.
It's better to keep the autocommands close to the domain they operate
in, otherwise you forget about them when working with that domain's
code, leading to unpleasant surpises.

While I was at it, I extracted the cursor stuff into its own plugin,
since it doesn't need to be lumped into this one.