cmdline keybindings, : ;
Though I want to use semicolon to enter cmdline mode, and I still want
access to semicolon's original functionality (hop to next occurence of
character on same line after a f/F/t/T command), I found that a straight
swap of colon and semicolon is a bad idea. In addition to doing the
wrong thing when I forget to use semicolon instead of colon, it does
the wrong thing in mappings that use colon. Using noremap is not always
an option (e.g. when leveraging a <Plug> map, which needs to be allowed
to remap to its intended command).
as a more convenient : and I still want access to
the functionality of ;