reimplement coherent#offset() with no redraw
Calling redraw was causing flicker as things like hlsearch highlights
were redrawn. This approach is more manual, but yields better results,
which is telling in terms of Vim's implementation and handling of
vimscript.

It was interesting getting this fine-tuned so the margins would behave
well as you approached the start/end of the buffer. Figuring out how
best to deal with small windows was also of interest.