Revert "change tag/pop modifier key"
This reverts commit 77a211137f82802a48dfa52f0dadd419c3a18484. My
original thinking on this was to use M-. and M-, in xmonad for
navigating to the next/prev workspace. But that conflicts with
Readline's M-. binding for yank-last-arg, which I make frequent use of
in the shell. There are alternative keys I could use for yank-last-arg,
namely M-_ or Esc . and I might revisit this decision, but for now, go
back to using M-. in applications and C-. in xmonad.