fixup Ag fzf: --nth 4..
We're using : as the delimiter, since ag output has the form,
  file:line:column:contents

But if contents happens to contain a : character, we don't want to limit
the search to the portion that precedes the colon -- rather, we want
to search all of contents. Ensure we do so by changing --nth to cover
everything from the 4th to the last field.